refactor(agent-api): split semantic scholar helpers
